\name{html}
\alias{myhtml}
\alias{getHtml}
\title{HTML functions for MACAT.}
\description{
 HTML functions for internal usage by other MACAT functions.
 Normally not called by user.
}
\details{
 Internal HTML functions. Not called by user.
}
\author{MACAT development team}
\seealso{\code{\link{plot.MACATevalScoring}}}
\keyword{internal}% at least one, from doc/KEYWORDS
