\name{Auxiliary Computation Functions}
\alias{squaredDist}
\alias{maxDistances}
\alias{pair.distances}
\alias{toPython}
\alias{rawDataToPython}
\alias{saveForPython}
\alias{slidingAverageAt}
\alias{compare.gammas}
\alias{getExpressionByChromosome.MACATData}
\alias{getExpressionByClass.MACATData}
\alias{getExpression.MACATData}
\alias{getExpressionByProbeset.MACATData}
\alias{demo.part}
\alias{crossvalParam}
\alias{evalParams}
\title{Auxiliary Functions for Computations in MACAT}
\description{
  Auxiliary functions for internal usage by other MACAT functions.
  Normally not called by user.
}
\details{
  Internal auxiliary functions. Not called by user.
}
\author{MACAT development team}
\seealso{\code{\link{evalScoring}}}
\keyword{internal}% at least one, from doc/KEYWORDS
