\name{html}
\alias{myhtml}
\alias{get.html}


%- Also NEED an '\alias' for EACH other topic documented here.
\title{HTML functions for MACAT.}
\description{
 HTML functions for internal usage by other MACAT functions.
 Normally not called by user.
}

\details{
Internal HTML functions. Not called by user.
}
%\value{
%  ~Describe the value returned
%}
%\references{ ~put references to the literature/web site here ~ }
\author{MACAT development team}
%\note{ ~~further notes~~ }
\seealso{\code{\link{plot.MACATevalScoring}}}
%\examples{}
\keyword{internal}% at least one, from doc/KEYWORDS
